Move the sources from lib to src
Change-type: patch
