+++ src/index.browser.ts
