+++ src/stores/virtual-storage.ts
