+++ tsconfig.dev.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "include": [
-    "lib/**/*",
+    "src/**/*",
     "tests/**/*"
   ]
 }
