+++ tsconfig.json
@@ -15,6 +15,6 @@
     "declaration": true
   },
   "include": [
-    "lib/**/*"
+    "src/**/*"
   ]
 }
